+++ extras/TestConceptSetCreation.R
@@ -36,12 +36,22 @@ vocabDatabaseSchema <- "merative_mdcr.cdm_merative_mdcr_v3788"
 
 
 conceptSets <- generateKeeperConceptSets(
-  phenotype = "Type I Diabetes Mellitus (T1DM)",
+  phenotype = "Acute liver failure",
+  clinicalDefinition = "Acute liver failure is a rare but life-threatening syndrome characterized by the rapid deterioration of hepatocellular function, manifesting as significant coagulopathy and hepatic encephalopathy of any grade, developing within 28 days of the onset of jaundice or initial hepatic symptoms in an individual without evidence of pre-existing chronic liver disease or cirrhosis. The syndrome arises from a direct, primary insult to hepatocytes — including viral, toxic, drug-induced, autoimmune, metabolic, or indeterminate causes — and is conceptually distinct from liver dysfunction occurring as a secondary consequence of hemodynamic compromise (e.g., ischemic hepatitis, shock liver), systemic sepsis, or passive hepatic congestion from right-sided heart failure, all of which are explicitly excluded.",
   client = client,
   vocabConnectionDetails = vocabConnectionDetails,
   vocabDatabaseSchema = vocabDatabaseSchema
 )
-readr::write_csv(conceptSets, "inst/t1dmConceptSets.csv")
+readr::write_csv(conceptSets, "e:/temp/conceptSets_ALF.csv")
+
+
+conceptSets <- generateKeeperConceptSets(
+  phenotype = "Type 2 diabetes",
+  client = client,
+  vocabConnectionDetails = vocabConnectionDetails,
+  vocabDatabaseSchema = vocabDatabaseSchema
+)
+readr::write_csv(conceptSets, "e:/temp/conceptSets_T2DM.csv")
 
 # conceptSetsOld <- readr::read_csv("e:/temp/afConceptSetsOld.csv")
 #